+++ content/developer-tools/migration-guides/java.mdx
@@ -149,7 +149,7 @@ WorkflowTriggerResult result = client.workflows().trigger(workflowTrigger);
 **New SDK:**
 
 ```java
-`import app.knock.api.core.JsonValue;
+import app.knock.api.core.JsonValue;
 import app.knock.api.models.workflows.WorkflowTriggerParams;
 import app.knock.api.models.workflows.WorkflowTriggerResponse;
