--- content/playground.mdx
@@ -1,99 +1,0 @@
----
-title: Playground
-description: A sample page to showcase available styles, components etc.
-Section: None
----
-
-## Typography
-
-# Heading 1
-
-## Heading 2
-
-### Heading 3
-
-#### Heading 4
-
-Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-
-> This is a block quote
-
-- An
-- Unordered
-- List
-
-1. A
-2. Ordered
-3. List
-
-## Components
-
-### `Callout`
-
-<Callout emoji="💡" title="Note" text={<>This is a callout component</>} />
-
-### `Attributes`
-
-<Attributes>
-  <Attribute
-    name="Attribute"
-    type="string"
-    description="This is an `Attribute` tag inside of an `Atributes` group."
-  />
-</Attributes>
-
-### `Accordion`
-
-<AccordionGroup>
-  <Accordion title="The first accordion item" description="They can also have little descriptions ✨">
-    This is the content inside of an accordion group.
-  </Accordion>
-  <Accordion title="The second accordion item">
-    This is the content inside of an accordion group.
-
-    ```java
-    class HelloWorld {
-        public static void main(String[] args) {
-            System.out.println("Hello, World!");
-        }
-    }
-    ```
-
-  </Accordion>
-  <Accordion title="The third accordion item">
-    This is the content inside of an accordion group.
-  </Accordion>
-</AccordionGroup>
-
-### `Step`
-
-<Steps>
-  <Step title="First Step">
-    These are instructions or content that only pertain to the first step.
-  </Step>
-  <Step title="Second Step">
-    These are instructions or content that only pertain to the second step.
-  </Step>
-  <Step title="Third Step">
-    These are instructions or content that only pertain to the third step.
-  </Step>
-</Steps>
-
-### `Card`
-
-Use `linkUrl` to give the card a clickthrough destination, and the `isExternal?` prop to determine whether it will open in a new tab
-
-<CardGroup>
-  <Card title="First card" emoji="✨" linkUrl="/something">
-    Content for the first card.
-  </Card>
-  <Card title="Second card" linkUrl="https://knock.app" isExternal={true}>
-    Content for the second card. This one opens in a new tab!
-  </Card>
-</CardGroup>
-
-### `MultiLangCodeBlock`
-
-The `MultiLangCodeBlock` will populate a switcher based on the examples provided in the snippet file.
-
-<MultiLangCodeBlock snippet="workflows.playground" title="Trigger a workflow" />
