--- data/integrationsSidebar.ts
@@ -1,6 +1,0 @@
-import { SidebarSection } from "./types";
-import integrationSidebarJson from "./integrationsSidebar.json";
-
-const sidebarContent: SidebarSection[] = integrationSidebarJson;
-
-export default sidebarContent;
