+++ data/sidebars/platformSidebar.ts
@@ -1,6 +1,6 @@
-import { SidebarSection } from "./types";
+import { SidebarSection } from "../types";
 
-const sidebarContent: SidebarSection[] = [
+export const PLATFORM_SIDEBAR: SidebarSection[] = [
   {
     title: "Getting started",
     slug: "/getting-started",
@@ -149,195 +149,3 @@ const sidebarContent: SidebarSection[] = [
     ],
   },
 ];
-
-const developerToolsContent: SidebarSection[] = [
-  {
-    title: "Developer tools",
-    slug: "/developer-tools",
-    desc: "Use our powerful developer tools in order to integrate Knock seamlessly into your development workflow.",
-    sidebarMenuDefaultOpen: true,
-    pages: [
-      { slug: "/overview", title: "Overview" },
-      { slug: "/api-keys", title: "API keys" },
-      { slug: "/service-tokens", title: "Service tokens" },
-      { slug: "/knock-cli", title: "Knock CLI" },
-      { slug: "/sdks", title: "SDKs" },
-      { slug: "/management-api", title: "Management API" },
-      { slug: "/api-logs", title: "API logs" },
-      { slug: "/knock-and-postman", title: "Knock and Postman" },
-      { slug: "/security", title: "Security" },
-      { slug: "/integrating-into-cicd", title: "Integrating into CI/CD" },
-      {
-        slug: "/outbound-webhooks",
-        title: "Outbound webhooks",
-        pages: [
-          { slug: "/overview", title: "Overview" },
-          { slug: "/event-types", title: "Event types" },
-        ],
-      },
-      { slug: "/validating-trigger-data", title: "Validating trigger data" },
-      {
-        slug: "/migration-guides",
-        title: "Migration guides",
-        pages: [
-          { slug: "/node", title: "Node.js 1.0" },
-          { slug: "/python", title: "Python 1.0" },
-          { slug: "/java", title: "Java 1.0" },
-          { slug: "/ruby", title: "Ruby 1.0" },
-          { slug: "/go", title: "Go 1.0" },
-        ],
-      },
-      {
-        slug: "/agent-toolkit",
-        title: "Agent Toolkit",
-        isBeta: true,
-        pages: [
-          { slug: "/overview", title: "Overview" },
-          { slug: "/getting-started", title: "Getting started" },
-          {
-            slug: "/workflows-as-tools",
-            title: "Workflows-as-tools",
-          },
-          {
-            slug: "/human-in-the-loop-flows",
-            title: "Human-in-the-loop flows",
-          },
-          {
-            slug: "/providing-context",
-            title: "Providing context",
-          },
-          {
-            slug: "/working-with-environments",
-            title: "Working with environments",
-          },
-          {
-            slug: "/tools-reference",
-            title: "Tools reference",
-          },
-        ],
-      },
-      {
-        slug: "/mcp-server",
-        title: "MCP server",
-        isBeta: true,
-      },
-      {
-        slug: "/building-with-llms",
-        title: "Building with LLMs",
-      },
-    ],
-  },
-];
-
-const tutorialsContent: SidebarSection[] = [
-  {
-    title: "Tutorials",
-    slug: "/tutorials",
-    sidebarMenuDefaultOpen: true,
-    pages: [
-      {
-        slug: "/overview",
-        title: "Overview",
-      },
-      {
-        slug: "/implementation-guide",
-        title: "Knock implementation guide",
-      },
-      {
-        slug: "/alerting",
-        title: "Alerting",
-      },
-      {
-        slug: "/customer-webhooks",
-        title: "Customer-facing webhooks",
-      },
-      {
-        slug: "/building-recurring-digests",
-        title: "Recurring digests",
-      },
-      { slug: "/migrate-from-courier", title: "Migrate from Courier" },
-      {
-        slug: "/modeling-users-objects-and-tenants",
-        title: "Modeling Users, Objects, and Tenants",
-      },
-    ],
-  },
-];
-
-const cliContent: SidebarSection[] = [
-  {
-    title: "Getting started",
-    slug: "/cli/overview",
-    pages: [
-      { slug: "", title: "Introduction" },
-      { slug: "/installation", title: "Install the Knock CLI" },
-      { slug: "/authentication", title: "Authentication" },
-      { slug: "/global-flags", title: "Global flags" },
-    ],
-    sidebarMenuDefaultOpen: true,
-  },
-  {
-    title: "Workflows",
-    slug: "/cli/workflow",
-    pages: [
-      { slug: "/list", title: "List workflows" },
-      { slug: "/get", title: "Get workflows" },
-      { slug: "/pull", title: "Pull workflows" },
-      { slug: "/push", title: "Push workflows" },
-      { slug: "/run", title: "Run workflow" },
-      { slug: "/validate", title: "Validate workflow" },
-      { slug: "/activate", title: "Activate workflow" },
-    ],
-  },
-
-  {
-    title: "Email layouts",
-    slug: "/cli/email-layout",
-    pages: [
-      { slug: "/list", title: "List email layouts" },
-      { slug: "/get", title: "Get email layouts" },
-      { slug: "/pull", title: "Pull email layouts" },
-      { slug: "/push", title: "Push email layouts" },
-      { slug: "/validate", title: "Validate email layouts" },
-    ],
-  },
-
-  {
-    title: "Translations",
-    slug: "/cli/translation",
-    pages: [
-      { slug: "/list", title: "List translations" },
-      { slug: "/get", title: "Get translations" },
-      { slug: "/pull", title: "Pull translations" },
-      { slug: "/push", title: "Push translations" },
-      { slug: "/validate", title: "Validate translations" },
-    ],
-  },
-
-  {
-    title: "Partials",
-    slug: "/cli/partial",
-    pages: [
-      { slug: "/list", title: "List partials" },
-      { slug: "/get", title: "Get partials" },
-      { slug: "/pull", title: "Pull partials" },
-      { slug: "/push", title: "Push partials" },
-      { slug: "/validate", title: "Validate partials" },
-    ],
-  },
-
-  {
-    title: "Commits",
-    slug: "/cli/commit",
-    pages: [
-      { slug: "/list", title: "List commits" },
-      { slug: "/get", title: "Get commits" },
-      { slug: "/all", title: "Commit changes" },
-      { slug: "/promote", title: "Promote changes" },
-    ],
-  },
-];
-
-export { tutorialsContent, developerToolsContent, cliContent };
-
-export default sidebarContent;
