+++ hooks/useSidebarContent.ts
@@ -1,14 +1,15 @@
 import { useRouter } from "next/router";
-import sidebarContent from "../data/sidebar";
-import { tutorialsContent, developerToolsContent } from "../data/sidebar";
+import { PLATFORM_SIDEBAR } from "@/data/sidebars/platformSidebar";
+import { TUTORIALS_SIDEBAR } from "@/data/sidebars/tutorialsSidebar";
+import { DEVELOPER_TOOLS_SIDEBAR } from "@/data/sidebars/developerToolsSidebar";
 
 export function useSidebarContent() {
   const { asPath } = useRouter();
   if (asPath.startsWith("/tutorials")) {
-    return tutorialsContent;
+    return TUTORIALS_SIDEBAR;
   }
   if (asPath.startsWith("/developer-tools")) {
-    return developerToolsContent;
+    return DEVELOPER_TOOLS_SIDEBAR;
   }
-  return sidebarContent;
+  return PLATFORM_SIDEBAR;
 }
