+++ layouts/MDXLayout.tsx
@@ -4,7 +4,6 @@ import { FrontMatter } from "../types";
 import { DocsLayout } from "./DocsLayout";
 import { IntegrationsLayout } from "./IntegrationsLayout";
 import { CliReferenceLayout } from "./CliReferenceLayout";
-import { ApiReferenceLayout } from "./ApiReferenceLayout";
 import InAppUILayout from "./InAppUILayout";
 
 import { useRouter } from "next/router";
@@ -15,10 +14,6 @@ const MDXLayout: React.FC<{ frontMatter: FrontMatter; sourcePath: string }> = (
   const { asPath } = useRouter();
 
   switch (props.frontMatter.layout) {
-    case "api":
-      return (
-        <ApiReferenceLayout {...props}>{props.children}</ApiReferenceLayout>
-      );
     case "integrations":
       return (
         <IntegrationsLayout {...props}>{props.children}</IntegrationsLayout>
