+++ package.json
@@ -13,7 +13,8 @@
     "format": "prettier \"**/*.+(ts|js|tsx|mdx|md)\"",
     "format.write": "yarn run format --write",
     "format.check": "yarn run format --check",
-    "generate-llms": "node scripts/generateLlmsTxt.mjs",
+    "generate-llms": "yarn run open-api-to-md && npx tsx scripts/generateLlmsTxt.ts",
+    "open-api-to-md": "bash scripts/openApiToMd.sh",
     "predev": "yarn generate-llms",
     "prebuild": "yarn generate-llms"
   },
@@ -82,12 +83,14 @@
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^7.31.0",
     "next-remote-refresh": "^0.10.0",
+    "openapi-to-md": "^1.0.25",
     "postcss": "^8.4.31",
     "prettier": "^2.3.2",
     "rehype-autolink-headings": "^7.0.0",
     "remark-slug": "^6.0.0",
     "tailwindcss": "^3.3.3",
     "ts-node": "^10.9.2",
+    "tsx": "^4.19.4",
     "typescript": "^5.2.2"
   }
 }
