+++ pages/api-reference/index.tsx
@@ -7,46 +7,11 @@ import remarkGfm from "remark-gfm";
 import { readOpenApiSpec, readStainlessSpec } from "@/lib/openApiSpec";
 import { CONTENT_DIR } from "@/lib/content.server";
 import { MDX_COMPONENTS } from "@/lib/mdxComponents";
-import { SidebarSection } from "@/data/types";
 import ApiReference from "@/components/ui/ApiReference/ApiReference";
-
-export const RESOURCE_ORDER = [
-  "workflows",
-  "messages",
-  "channels",
-  "users",
-  "objects",
-  "tenants",
-  "recipients",
-  "schedules",
-  "audiences",
-  "bulk_operations",
-  "providers",
-  "shared",
-];
-
-export const PRE_SIDEBAR_CONTENT: SidebarSection[] = [
-  {
-    title: "API reference",
-    slug: `/api-reference/overview`,
-    pages: [
-      { slug: "/", title: "Overview" },
-      { slug: "/client-libraries", title: "Client libraries" },
-      { slug: "/api-keys", title: "API keys" },
-      { slug: "/authentication", title: "Authentication" },
-      { slug: "/rate-limits", title: "Rate limits" },
-      { slug: "/batch-rate-limits", title: "Batch rate limits" },
-      { slug: "/idempotent-requests", title: "Idempotent requests" },
-      { slug: "/data-retention", title: "Data retention" },
-      { slug: "/bulk-endpoints", title: "Bulk endpoints" },
-      { slug: "/trigger-data-filtering", title: "Trigger data filtering" },
-      { slug: "/pagination", title: "Pagination" },
-      { slug: "/errors", title: "Errors" },
-      { slug: "/error-codes", title: "Common error codes" },
-    ],
-    sidebarMenuDefaultOpen: true,
-  },
-];
+import {
+  RESOURCE_ORDER,
+  API_REFERENCE_OVERVIEW_CONTENT,
+} from "@/data/sidebars/apiOverviewSidebar";
 
 function ApiReferencePage({ openApiSpec, stainlessSpec, preContentMdx }) {
   return (
@@ -56,7 +21,7 @@ function ApiReferencePage({ openApiSpec, stainlessSpec, preContentMdx }) {
       stainlessSpec={stainlessSpec}
       preContent={<MDXRemote {...preContentMdx} components={MDX_COMPONENTS} />}
       resourceOrder={RESOURCE_ORDER}
-      preSidebarContent={PRE_SIDEBAR_CONTENT}
+      preSidebarContent={API_REFERENCE_OVERVIEW_CONTENT}
     />
   );
 }
